move code to be cleaned to legacy package
