+++ web/app/cad/legacy/mesh/workbench.js
@@ -2,7 +2,7 @@ import Vector from 'math/vector';
 import * as cad_utils from '../../cad-utils'
 import * as math from '../../../math/math'
 import {HashTable} from '../../../utils/hashmap'
-import {Mesh} from '../../mesh'
+import {Mesh} from '../mesh'
 import revolve from './revolve'
 import {Triangulate} from '../../tess/triangulation'
